L4Re target spec: fail-fast + better library location code
GCC is mandatory for this target, that might need documentation
somewhere. As soon as something mandatory cannot be found, the compiler
will panic. This is intended, because the functions involved don't allow
the usage of a Result type.
libgcc_eh is now passed using "-l" and crtbeginT.o and crtend.o are now
located using `gcc -print-filename`.

Co-authored-by: Tobias Schaffner <tschaff@genua.de>